+++ util/class.tx_rnbase_util_TCA.php
@@ -57,6 +57,18 @@ class tx_rnbase_util_TCA {
 		}
 		return $GLOBALS['TCA'][$tableName]['ctrl']['languageField'];
 	}
+	/**
+	 * Liefert den Spaltennamen für den Titel der Tabelle.
+	 *
+	 * @param string $tableName
+	 * @return string
+	 */
+	public static function getLabelFieldForTable($tableName) {
+		if (empty($GLOBALS['TCA'][$tableName]) || empty($GLOBALS['TCA'][$tableName]['ctrl']['label'])) {
+			return '';
+		}
+		return $GLOBALS['TCA'][$tableName]['ctrl']['label'];
+	}
 
 	/**
 	 * Load TCA for a specific table. Since T3 6.1 the complete TCA is loaded.